requirements.txt
asgiref==3.5.0
boto3==1.23.0
botocore==1.26.1
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==21.1
django-friendship==1.9.6
django-heroku==0.3.1
django-rest-friendship==0.2.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.0
response==0.5.0
s3transfer==0.5.2
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
